+++ src/aks-preview/azext_aks_preview/_consts.py
@@ -64,6 +64,7 @@ CONST_LOAD_BALANCER_SKU_STANDARD = "standard"
 # ManagedClusterSKU Tier
 CONST_MANAGED_CLUSTER_SKU_TIER_FREE = "free"
 CONST_MANAGED_CLUSTER_SKU_TIER_STANDARD = "standard"
+CONST_MANAGED_CLUSTER_SKU_TIER_PREMIUM = "premium"
 
 CONST_OUTBOUND_MIGRATION_MULTIZONE_TO_NATGATEWAY_MSG = "Warning: this AKS cluster has multi-zonal nodepools, but NAT Gateway is not currently zone redundant. Migrating outbound connectivity to NAT Gateway could lead to a reduction in zone redundancy for this cluster. Continue?"
 # load balancer backend pool type