+++ src/aks-preview/HISTORY.rst
@@ -12,10 +12,17 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
-0.5.161
+0.5.162
 +++++++
 * Add `get-upgrades` and `get-revisions` to the `az aks mesh` command.
 * Add `az aks mesh upgrade` commands to manage upgrades for Azure Service Mesh.
+
+0.5.161
++++++++
+* Support `premium` cluster sku tier in `az aks create` and `az aks update` commands
+* Add option `--k8s-support-plan` to `az aks create` and `az aks update` commands
+* Add `az aks machine list` command to fetch list of machines in an agentpool.
+* Add `az aks machine show` command to fetch information about a specific machine in an agentpool.
 
 0.5.160
 +++++++