+++ src/fleet/HISTORY.rst
@@ -31,3 +31,17 @@ Release History
 0.2.4
 ++++++
 * Added support for private fleet & MSI.
+
+0.2.5
+++++++
+* Upgrade SDK version to 2023-08-15-preview
+
+0.2.6
+++++++
+* By default, fleets are now created without a hub
+* Added support for Fleet Upgrade Strategy
+* Added argument for Node Image Selection.
+
+0.2.7
+++++++
+* Fix for `az fleet updaterun --node-image-selection` argument.